APPT-1329: GovNotify retries (#1078)
# Description

[Ticket 1329](https://nhsd-jira.digital.nhs.uk/browse/APPT-1329) claims
that when we receive a 400 error from GovNotify we still retry the
request all 3 times, and that this is waste because we know the 2nd &
3rd attempts will fail. It says that retries should only be attempted
when we receive a 429.

As far as I can tell, this is already how this code behaves. This PR
simply adds tests to prove this behaviour. [The docs on
this](https://docs.notifications.service.gov.uk/rest-api.html#errors)
are extremely bad and vague about what exactly the API returns, so there
is a chance that we're getting a different result back from real
GovNotify. I've added a unit testing proving that their example error
message in the docs gets parsed correctly, and also logged out the
status code in the logs. If behaviour differs in a live environment this
should alert us to it.

Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests