APPT-1489: Cancellation notification status (#1089)
# Description

When we introduced the CancellationNotificationStatus enum, I think it
was with the intention of tracking when clinicians have manually
contacted an individual. It's been raised as a bug that we're not
setting it when we attempt automatic cancellations, so this PR adds that
in.
<img width="818" height="326" alt="image"
src="https://github.com/user-attachments/assets/df902e21-532a-4b5c-9804-a5d465864aa2"
/>

Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests