APPT-1573: Edit region bug (#1120)
# Description

Region and ICB users get access to sites which belong to that
region/icb. They have the permission to change that site's ICB/Region,
which means they lose access to the site in the process.

They need to retain the ability to do this, but rather than being
redirected to the site details page after editing the details they need
to be redirected to the site selection page to avoid 403 errors.

It would seem that the site cache isn't currently being hit by the
FetchPermissions route, but if it ever was then we'd likely create a bug
in this area. If this happens, we'd need to invalidate the cache at the
same time as applying the region/icb code edit to the site.

Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests
- [ ] If I've added/updated an end-point, I've added the appropriate
annotations and tested the Swagger documentation reflects the change