APPT-1586: Fix caching bug in GetAllSites (#1131)
# Description

There is currently a bug in the GetAllSites method of the SiteService.
We cache the result regardless of whether `includeDeleted` was true or
false. This means if you call it once with false you'll get only the
live sites, then if you call it a 2nd time with true you would expect
the live and deleted sites but you'll instead get the result of the
cache which is only the live sites.

The fix for this is to always cache the entire list and filter it down
in memory when deleted sites should be omitted.

Fixes # ([issue](https://nhsd-jira.digital.nhs.uk/browse/APPT-1586))

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests
- [ ] If I've added/updated an end-point, I've added the appropriate
annotations and tested the Swagger documentation reflects the change