APPT-XXX: Only check for siteId when the optional parameter is provided (#1152)
# Description


[APPT-1470](https://github.com/NHSDigital/nbs-appointments-management-service/pull/1139)
introduced a check for the site's exisence when cancelling a booking,
and returned NotFound if no matching site was found. The problem is,
site is an optional parameter in the CancelBooking endpoint. This causes
a BadRequest exception from Cosmos which apparently doesn't let you pass
the empty string as the id 🤷

It's a bit worrying this wasn't caught in tests, so I have a TODO to add
some. Potentially in a 2nd PR if we need to get this in soon.

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests
- [ ] If I've added/updated an end-point, I've added the appropriate
annotations and tested the Swagger documentation reflects the change