APPT-1613: Only check for siteId when the optional parameter is provid… (#1153)
Cherry-pick of 5f38fa5e2cc16d10b1b583062c3e0ac8b236b7a4 into 2.8:


[APPT-1470](https://github.com/NHSDigital/nbs-appointments-management-service/pull/1139)
introduced a check for the site's exisence when cancelling a booking,
and returned NotFound if no matching site was found. The problem is,
site is an optional parameter in the CancelBooking endpoint. This causes
a BadRequest exception from Cosmos which apparently doesn't let you pass
the empty string as the id 🤷

It's a bit worrying this wasn't caught in tests, so I have a TODO to add
some. Potentially in a 2nd PR if we need to get this in soon.