APPT-1597: Thread safe updates to site cache (#1157)
# Description

We don't currently update the site cache when we edit site details,
including when we edit region (the bug in
[APPT-1597](https://nhsd-jira.digital.nhs.uk/browse/APPT-1597)). We need
to refresh the cache whenever a site is changed or else we'll return
incorrect cached data.

We could simply remove it, but with over 8,000 sites in the system
that's costly. It would be better to patch it. A patch however
introduces race conditions, as two parallel edits to different sites
could both be trying to edit the cache at the same time.

This PR uses
[semaphores](https://learn.microsoft.com/en-us/dotnet/api/system.threading.semaphoreslim?view=net-9.0)
to update this cache in a thread-safe way, and adds tests to assert that
every method in `SiteService` which updates a site also updates the
cache if the operation was successful.

Fixes # ([issue](https://nhsd-jira.digital.nhs.uk/browse/APPT-1597))

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests
- [ ] If I've added/updated an end-point, I've added the appropriate
annotations and tested the Swagger documentation reflects the change