APPT-1670: Decode query string prams in the Postman pre-request script (#1178)
# Description

When making requests to our backend using our Postman collection, we
authenticate by signing the requests with an HMAC key. When the backend
receives a request, it generates what it expects the signature to be
then compares the signature it actually received. One of the inputs to
generating this signature is the query string of the request.

The backend is decoding the query string before passing it into SHA to
generate the expected signature, but the Postman pre-request script does
not. Therefore if you use characters which require URL encoding in a
Postman parameter, you'll incorrectly receive a 401 response.

The fix for this is simply to instruct Postman to encode the query
string before signing its requests.

Fixes # ([issue](https://nhsd-jira.digital.nhs.uk/browse/APPT-1670))

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests
- [ ] If I've added/updated an end-point, I've added the appropriate
annotations and tested the Swagger documentation reflects the change