APPT-1526: Stop calling redirect within a trycatch block (#1187)
# Description

NextJS intentionally throws an internal error when `redirect()` is
called, but we're wrapping all of our calls to the appointment service
in a trycatch. This means that redirects for auth expiry are being
caught and logged as errors.

We introduced the concept of a ServerActionResult because exceptions
thrown within declarative server actions weren't triggering NextJS'
error boundaries. During the solve for this work I've discovered that if
everything is handled with promise.reject() instead of throwing
exceptions, the boundary is correctly invoked. Any genuine `Error`s
thrown by server actions get passed to the onReject callback as errors.

The complexity in this work is caused by us wanting to do different
things in different situations. On 403 we want to throw a special type
of error, on 404 we want to invoke one of NextJS' methods, on 401 we
want to delete the token then redirect.

The most sensible fix I've come up with is to introduce a
`ServerActionFailure` object. We can then guarantee that when a
`fromServer()` promise chain is rejected, the callback will be invoked
either by an `Error` or a `ServerActionFailure`. If it's an error we
throw and invoke the error boundary, but if it's our failure class we
then invoke its `handle()` method. To achieve this, we need to stop
redirecting directly from `appointmentsService` itself.

Fixes # (issue)

# Checklist:

- [ ] My work is behind a feature toggle (if appropriate)
- [ ] If my work is behind a feature toggle, I've added a full suite of
tests for both the ON and OFF state
- [ ] The ticket number is in the Pull Request title, with format
"APPT-XXX: My Title Here"
- [ ] I have ran npm tsc / lint (in the future these will be ran
automatically)
- [ ] My code generates no new .NET warnings (in the future these will
be treated as errors)
- [ ] If I've added a new Function, it is disabled in all but one of the
terraform groups (e.g. http_functions)
- [ ] If I've added a new Function, it has both unit and integration
tests. Any request body validators have unit tests also
- [ ] If I've made UI changes, I've added appropriate Playwright and
Jest tests
- [ ] If I've added/updated an end-point, I've added the appropriate
annotations and tested the Swagger documentation reflects the change